book_service/urls.py
from django.urls import path, include
from rest_framework import routers

from book_service.views import BookViewSet

router = routers.DefaultRouter()
router.register("", BookViewSet, basename="books")

urlpatterns = [path("", include(router.urls))]

app_name = "book_service"
