requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coverage==7.2.3
cron-descriptor==1.2.35
distlib==0.3.6
Django==4.2
django-celery-beat==2.5.0
django-rest-framework==0.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
filelock==3.11.0
flower==1.2.0
frozenlist==1.3.3
humanize==4.6.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
kombu==5.2.4
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.6.0
pyrsistent==0.19.3
pyTelegramBotAPI==4.11.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
telebot==0.0.5
tornado==6.3.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
virtualenv==20.21.0
wcwidth==0.2.6
yarl==1.8.2
